fix(tools): node_exec sub-second timeout_sec truncated to zero
time.Duration(0.05) is an integer conversion and truncates to 0, so any
timeout_sec below one second silently fell back to the 120s default
invoke timeout. Convert through float64 seconds instead. Caught by CI
(TestNodeExecToolTimeoutBounded).

+++ internal/tools/node_exec_tool.go
@@ -114,7 +114,10 @@ func (t *NodeExecTool) Execute(ctx context.Context, args map[string]any) *Result
 	res, invokeErr := nodes.Invoke(ctx, t.registry, t.nodeStore, node.ID, nodes.InvokeRequest{
 		Command: command,
 		Args:    execArgs,
-		Timeout: time.Duration(timeoutSec) * time.Second,
+		// float64 seconds → Duration must multiply through float64: an
+		// integer conversion (time.Duration(0.05)) truncates to 0, which
+		// Invoke would then replace with the default timeout.
+		Timeout: time.Duration(timeoutSec * float64(time.Second)),
 	})
 	t.auditInvoke(agentID, node, invokeErr)
